Update redirect paths to new '/tabs/home'
Changed the navigation redirect paths from '/home' to '/tabs/home' in auth.page.ts after successful login or register. This was done due to the reorganisation of the application structure, where the home component is now nested under the 'tabs' directory. The updating of the paths ensures correct redirection to the home page, improving the user navigation experience within the app. Associated routes in app-routing.module.ts are also modified to reflect these changes.